+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+- fixes exception when using custom site models
+
 
 `0.13`_ (2021-10-08)
 ~~~~~~~~~~~~~~~~~~~~